Hotfix incorrect arena reads
Return value of get_int, get_byte and get_short was 'int' but should
have been 'long' to match the size of a t_proc.reg.
Fixed that, updated tests to match.

+++ corewar/includes/arena.h
@@ -50,10 +50,10 @@ char				*g_arena;
 /* 'public api' */
 void			store_register_contents(t_proc *ps, int reg_num, int req_idx);
 unsigned long	bigendian_num(char *buf, size_t size);
-int				get_int(t_proc *ps, int req_idx);
-int				get_short(t_proc *ps, int req_idx);
-int				get_byte(t_proc *ps, int req_idx);
-int				follow_indirect_reference(t_proc *ps, int index_of_indirect);
+long			get_int(t_proc *ps, int req_idx);
+long			get_short(t_proc *ps, int req_idx);
+long			get_byte(t_proc *ps, int req_idx);
+long			follow_indirect_reference(t_proc *ps, int index_of_indirect);
 /* internals */
 void		_store_register_contents(t_proc *ps, int reg_idx, int req_idx);
 void		_lstore_register_contents(t_proc *ps, int reg_idx, int req_idx);
